CardBehavior를 만들어서 behavior 로직 분리
