+++ playingCard/playingCard/ViewController.swift
@@ -31,26 +31,9 @@ class ViewController: UIViewController {
             faceUpCardViews[0].suit == faceUpCardViews[1].suit
     }
     
-    lazy var animator: UIDynamicAnimator = {
-        return UIDynamicAnimator(referenceView: view)
-    }()
+    lazy var animator: UIDynamicAnimator = UIDynamicAnimator(referenceView: view)
     
-    lazy var collistionBehavior: UICollisionBehavior = {
-        let behavior = UICollisionBehavior()
-        behavior.translatesReferenceBoundsIntoBoundary = true
-        animator.addBehavior(behavior)
-        return behavior
-    }()
-    
-    lazy var itemBehavior: UIDynamicItemBehavior = {
-        let behavior = UIDynamicItemBehavior()
-        behavior.allowsRotation = false
-        // 1.0 = 가속도 x, 1.1 부딫힐수록 점점 빨라짐, 0.9 점점 느려짐
-        behavior.elasticity = 1.0
-        behavior.resistance = 0
-        animator.addBehavior(behavior)
-        return behavior
-    }()
+    lazy var cardBehavior: CardBehavior = CardBehavior(in: animator)
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -66,16 +49,7 @@ class ViewController: UIViewController {
             cardView.suit = card.suit.rawValue
             let tap = UITapGestureRecognizer(target: self, action: #selector(flipCard(_:)))
             cardView.addGestureRecognizer(tap)
-            collistionBehavior.addItem(cardView)
-            itemBehavior.addItem(cardView)
-            let push = UIPushBehavior(items: [cardView], mode: .instantaneous)
-            push.angle = (2*CGFloat.pi).arc4random
-            push.magnitude = CGFloat(1.0) + CGFloat(2.0).arc4random
-            // push.action이 실행되려면 반드시 push가 힙에 존재해야한다. 그래서 unowned 사용
-            push.action = { [unowned push] in
-                push.dynamicAnimator?.removeBehavior(push)
-            }
-            animator.addBehavior(push)
+            cardBehavior.addItem(cardView)
         }
     }
